requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.7
django-apscheduler==0.6.2
django-celery-beat==2.4.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
executing==1.2.0
flower==1.2.0
gunicorn==20.1.0
humanize==4.6.0
inflection==0.5.1
ipdb==0.13.11
ipython==8.11.0
jedi==0.18.2
jsonschema==4.17.3
kombu==5.2.4
matplotlib-inline==0.1.6
numpy==1.24.2
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
